+++ tests/utilities/test_auto_restart.py
@@ -16,6 +16,7 @@ import os
 import random
 from collections.abc import Iterable
 from typing import Optional
+from unittest import mock
 
 import numpy as np
 import pytest
@@ -28,10 +29,18 @@ from torch.utils.data.dataloader import DataLoader, default_collate
 from torch.utils.data.dataset import Dataset, IterableDataset
 
 import tests.helpers.utils as tutils
-from pytorch_lightning import seed_everything
-from pytorch_lightning.utilities.auto_restart import CaptureIterableDataset, FastForwardSampler
+from pytorch_lightning import seed_everything, Trainer
+from pytorch_lightning.trainer.supporters import CombinedLoader
+from pytorch_lightning.utilities.apply_func import apply_to_collection
+from pytorch_lightning.utilities.auto_restart import (
+    _dataloader_load_state_dict,
+    _dataloader_to_state_dict,
+    CaptureIterableDataset,
+    FastForwardSampler,
+)
 from pytorch_lightning.utilities.enums import AutoRestartBatchKeys
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
+from pytorch_lightning.utilities.imports import _fault_tolerant_enabled
 from tests.helpers.runif import RunIf
 
 
@@ -164,7 +173,7 @@ def test_fast_forward_on_random_sampler():
 
     dataset = range(15)
     generator = torch.Generator().manual_seed(seed)
-    values = [v for v in RandomSampler(dataset, generator=generator)]
+    values = list(RandomSampler(dataset, generator=generator))
 
     generator = torch.Generator().manual_seed(seed)
     random_sampler = RandomSampler(dataset, generator=generator)
@@ -205,12 +214,12 @@ def test_fast_forward_on_random_sampler():
 
 class RangeIterableDataset(IterableDataset):
 
-    def __init__(self, data, num_workers: int, batch_size: int, is_in_workers: bool, state_dict=None):
+    def __init__(self, data, num_workers: int, batch_size: int, state_dict=None, attr_name: str = "iter_sampler"):
         self.data = list(data)
         self.batch_size = batch_size
         self.num_workers = num_workers
-        self.is_in_workers = is_in_workers
         self.state_dict = state_dict
+        self.attr_name = attr_name
 
     def __iter__(self):
         worker_info = get_worker_info()
@@ -225,11 +234,12 @@ class RangeIterableDataset(IterableDataset):
         else:
             self.user_sampler = RandomSampler(self.data)
 
-        self.iter_sampler = iter(self.user_sampler)
+        setattr(self, self.attr_name, iter(self.user_sampler))
         return self
 
     def __next__(self):
-        return self.data[next(self.iter_sampler)]
+        iter_sampler = getattr(self, self.attr_name)
+        return self.data[next(iter_sampler)]
 
 
 @pytest.mark.skipif(torch.cuda.is_available(), reason="This test takes around 30 sec and should be skipped in Azure CI")
@@ -246,7 +256,10 @@ def test_fast_forward_sampler_over_iterative_dataset(num_workers):
     generator.manual_seed(initial_seed)
     dataset = RangeIterableDataset(range(20), num_workers, batch_size, True)
     dataset = CaptureIterableDataset(dataset, num_workers)
+
     dataloader = DataLoader(dataset, batch_size=batch_size, num_workers=num_workers, generator=generator)
+    Trainer._add_sampler_metadata_collate(dataloader)
+
     iter_dataloader = iter(dataloader)
     batches = []
     for _ in range(5):
@@ -256,18 +269,20 @@ def test_fast_forward_sampler_over_iterative_dataset(num_workers):
 
     state_dict = {'iter_sampler': {}}
     for batch in batches[:2]:
-        _state_dict = CaptureIterableDataset.convert_batch_into_state_dict(batch)
-        for k, v in _state_dict.items():
+        batch, _state_dict = CaptureIterableDataset.extract_samplers_state_dict_from_batch(batch)
+        for k, v in _state_dict[0].items():
             state_dict[k].update(v)
 
     assert len(state_dict["iter_sampler"]) == (num_workers if num_workers > 1 else 1)
 
     initial_seed = seed_everything(42)
     generator.manual_seed(initial_seed)
-    dataset = RangeIterableDataset(range(20), num_workers, batch_size, True, state_dict=state_dict)
+    dataset = RangeIterableDataset(range(20), num_workers, batch_size, state_dict=state_dict)
     dataset = CaptureIterableDataset(dataset)
     dataset.load_state_dict(state_dict)
     dataloader = DataLoader(dataset, batch_size=batch_size, num_workers=num_workers, generator=generator)
+    Trainer._add_sampler_metadata_collate(dataloader)
+
     iter_dataloader = iter(dataloader)
     batches_restart = []
     for _ in range(3):
@@ -362,7 +377,7 @@ class MetaLearningDataset(IterableDataset):
         num_workers: Optional[int] = None,
         global_rank: Optional[int] = None,
         world_size: Optional[int] = None,
-        initial_seed: Optional[torch.Generator] = None,
+        initial_seed: Optional[int] = None,
         shuffle: bool = True,
         debugging: bool = False,
     ):
@@ -522,6 +537,7 @@ def _test_fast_forward_sampler_with_distributed_sampler_and_iterative_dataset(ra
     )
     dataset = CaptureIterableDataset(dataset, initial_seed=initial_seed)
     dataloader = DataLoader(dataset, num_workers=num_workers, batch_size=1, generator=generator)
+    Trainer._add_sampler_metadata_collate(dataloader)
 
     epoch_results = []
     for _ in range(2):
@@ -544,8 +560,8 @@ def _test_fast_forward_sampler_with_distributed_sampler_and_iterative_dataset(ra
         assert torch.equal(
             epoch_results[0][0]["data"]["selected_indexes"], epoch_results[0][1]["data"]["selected_indexes"]
         )
-        assert epoch_results[0][2][AutoRestartBatchKeys.PL_SAMPLERS]["id"] == 0
-        assert epoch_results[0][3][AutoRestartBatchKeys.PL_SAMPLERS]["id"] == 1
+        assert 0 in epoch_results[0][2][AutoRestartBatchKeys.PL_SAMPLERS]["iter_sampler"]  # worker id 0
+        assert 1 in epoch_results[0][3][AutoRestartBatchKeys.PL_SAMPLERS]["iter_sampler"]  # worker id 1
         assert not torch.equal(epoch_results[0][2]["data"][0], epoch_results[0][3]["data"][0])
     else:
         first_task_metadata = all_gather(epoch_results[0][0]["data"]["task_length"], worldsize)
@@ -562,8 +578,8 @@ def _test_fast_forward_sampler_with_distributed_sampler_and_iterative_dataset(ra
     # restarting on epoch 0 / real batch 2
     state_dict = {'iter_sampler': {}}
     for batch in epoch_results[0][2:4]:
-        metadata = CaptureIterableDataset.convert_batch_into_state_dict(batch)
-        for k, v in metadata.items():
+        batch, _state_dict = CaptureIterableDataset.extract_samplers_state_dict_from_batch(batch)
+        for k, v in _state_dict[0].items():
             state_dict[k].update(v)
 
     dataset = ClassificationDataset(range(dataset_length), labels)
@@ -582,6 +598,7 @@ def _test_fast_forward_sampler_with_distributed_sampler_and_iterative_dataset(ra
     dataset = CaptureIterableDataset(dataset, initial_seed=initial_seed)
     dataset.load_state_dict(state_dict)
     dataloader = DataLoader(dataset, num_workers=num_workers, batch_size=1, generator=generator)
+    Trainer._add_sampler_metadata_collate(dataloader)
 
     epoch_results_restart = []
     for _ in range(2):
@@ -625,3 +642,216 @@ def test_fast_forward_sampler_with_distributed_sampler_and_iterative_dataset():
     mp.spawn(
         _test_fast_forward_sampler_with_distributed_sampler_and_iterative_dataset, args=(worldsize, ), nprocs=worldsize
     )
+
+
+@mock.patch.dict(os.environ, {'PL_FAULT_TOLERANT_TRAINING': "1"})
+@RunIf(max_torch="1.6")
+def test_fault_tolerant_not_supported():
+    with pytest.raises(MisconfigurationException, match="Restart is only supported with torch >= 1.7.0."):
+        _fault_tolerant_enabled()
+
+
+def create_iterable_dataset(batch_size, num_workers, attr_name="iter_sampler"):
+    dataset = RangeIterableDataset(range(50), num_workers=num_workers, batch_size=batch_size, attr_name=attr_name)
+    return CaptureIterableDataset(dataset)
+
+
+def create_dataloader():
+    dataset = range(50)
+    num_workers = 2
+    batch_size = 8
+    sampler = FastForwardSampler(SequentialSampler(dataset))
+    sampler.setup(batch_size)
+
+    dataloader = DataLoader(
+        dataset,
+        sampler=sampler,
+        batch_size=batch_size,
+    )
+    dataloader.fast_forward_sampler = sampler
+
+    loader_dict = {
+        "a": [
+            DataLoader(create_iterable_dataset(3, num_workers), num_workers=num_workers, batch_size=3),
+            dataloader,
+        ],
+        "b": DataLoader(
+            create_iterable_dataset(2, num_workers=1, attr_name="custom_sampler"), num_workers=0, batch_size=2
+        )
+    }
+    apply_to_collection(loader_dict, DataLoader, Trainer._add_sampler_metadata_collate)
+    return CombinedLoader(loader_dict)
+
+
+# Lightning will wrap the iterator within a prefect function as follow.
+def prefetch_iterator(iterable: Iterable):
+    it = iter(iterable)
+
+    try:
+        # the iterator may be empty from the beginning
+        last = next(it)
+    except StopIteration:
+        return
+
+    for val in it:
+        # yield last and has next
+        yield last, False, it
+        last = val
+    # yield last, no longer has next
+    yield last, True, it
+
+
+@pytest.mark.skipif(torch.cuda.is_available(), reason="This test takes around 15 sec and should be skipped in Azure CI")
+@mock.patch.dict(os.environ, {"PL_FAULT_TOLERANT_TRAINING": "1"})
+@RunIf(min_torch="1.7.0")
+def test_combined_dataloader_state_dict_and_reload():
+    """
+    This test makes sure the CombinedLoader used in the condition of Lightning properly
+    capture its children DataLoader states.
+    """
+
+    dataloader = create_dataloader()
+
+    iter_dataloader = iter(prefetch_iterator(dataloader))
+    num_batches_processed = 4
+    for idx in range(1, num_batches_processed):
+        _, _, prefetched_iterator = next(iter_dataloader)
+
+        loader_iters = prefetched_iterator._loader_iters
+
+        # when dealing with IterativeDataset,
+        # the sampler state dict will be attached directly onto the iterator to simplify collection.
+
+        if idx == 1:
+            assert loader_iters["a"][0]._sampler_state_dict == [{'iter_sampler': {0: {'current_iteration': 3}}}]
+            assert loader_iters["a"][1]._sampler_state_dict == []
+            assert loader_iters["b"]._sampler_state_dict == [{'custom_sampler': {0: {'current_iteration': 2}}}]
+        elif idx == 2:
+            assert loader_iters["a"][0]._sampler_state_dict == [{
+                'iter_sampler': {
+                    0: dict(current_iteration=3),
+                    1: dict(current_iteration=3)
+                }
+            }]
+            assert loader_iters["a"][1]._sampler_state_dict == []
+            assert loader_iters["b"]._sampler_state_dict == [{'custom_sampler': {0: {'current_iteration': 4}}}]
+        else:
+            assert loader_iters["a"][0]._sampler_state_dict == [{
+                'iter_sampler': {
+                    0: dict(current_iteration=6),
+                    1: dict(current_iteration=3)
+                }
+            }]
+            assert loader_iters["a"][1]._sampler_state_dict == []
+            assert loader_iters["b"]._sampler_state_dict == [{'custom_sampler': {0: {'current_iteration': 6}}}]
+
+    state_dict = dataloader.state_dict(num_batches_processed=3)
+
+    expected = {
+        "b": {
+            "num_workers": 0,
+            "previous_worker": None,
+            "custom_sampler": {
+                0: dict(current_iteration=6)
+            },
+        },
+        "a": [
+            {
+                "num_workers": 2,
+                "previous_worker": 1,
+                "iter_sampler": {
+                    0: dict(current_iteration=6),
+                    1: dict(current_iteration=3)
+                },
+            },
+            {
+                "num_workers": 0,
+                "previous_worker": None,
+                0: dict(current_iteration=24)
+            },
+        ],
+    }
+    assert state_dict == expected
+
+    dataloader = create_dataloader()
+    apply_to_collection(dataloader, DataLoader, Trainer._add_sampler_metadata_collate)
+    dataloader.load_state_dict(state_dict)
+
+    iter_dataloader = iter(prefetch_iterator(dataloader))
+    _, _, prefetched_iterator = next(iter_dataloader)
+
+    loader_iters = prefetched_iterator._loader_iters
+
+    assert loader_iters["a"][0]._sampler_state_dict == [{
+        'num_workers': 2,
+        'iter_sampler': {
+            0: dict(current_iteration=6),
+            1: dict(current_iteration=6)
+        }
+    }]
+    assert loader_iters["a"][1]._sampler_state_dict == []
+    assert loader_iters["b"]._sampler_state_dict == [{
+        'num_workers': 0,
+        'custom_sampler': {
+            0: dict(current_iteration=8)
+        }
+    }]
+
+    state_dict = dataloader.state_dict(num_batches_processed=4)
+
+    expected = {
+        "a": [
+            {
+                "num_workers": 2,
+                "previous_worker": 0,
+                "iter_sampler": {
+                    0: dict(current_iteration=6),
+                    1: dict(current_iteration=6)
+                },
+            },
+            {
+                "num_workers": 0,
+                "previous_worker": None,
+                0: dict(current_iteration=32)
+            },
+        ],
+        "b": {
+            "num_workers": 0,
+            "previous_worker": None,
+            "custom_sampler": {
+                0: dict(current_iteration=8)
+            },
+        },
+    }
+
+    assert state_dict == expected
+
+
+def test_dataloader_to_state_dict_and_reload():
+    """
+    Note: Those utilities are used only with DataLoader wrapping a ``mapping`` based dataset.
+    """
+
+    def create_dataloader():
+        dataset = range(50)
+        batch_size = 8
+        sampler = FastForwardSampler(SequentialSampler(dataset))
+        sampler.setup(batch_size)
+
+        return DataLoader(dataset, sampler=sampler, batch_size=batch_size)
+
+    dataloader = create_dataloader()
+    iter_dataloader = iter(dataloader)
+    _ = next(iter_dataloader)
+    _ = next(iter_dataloader)
+
+    state_dict = _dataloader_to_state_dict(dataloader, iter_dataloader)
+    assert state_dict == {'num_workers': 0, 'previous_worker': None, 0: {'current_iteration': 16}}
+
+    dataloader = create_dataloader()
+    dataloader = _dataloader_load_state_dict(dataloader, state_dict)
+    iter_dataloader = iter(dataloader)
+    _ = next(iter_dataloader)
+
+    state_dict = _dataloader_to_state_dict(dataloader, iter_dataloader)
+    assert state_dict == {'num_workers': 0, 'previous_worker': None, 0: {'current_iteration': 24}}